# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PR-мониторинг API",
    "description": "API веб-интерфейса команды: факты публикаций, перепечатки, охват, автоматический поиск упоминаний.\n\nАвтоматизация (n8n) использует эндпоинты `/api/cron/*` — они защищены заголовком `X-Automation-Api-Key`. Кабинет команды и админ-панель (`/api/team/*`, `/api/admin/*`) работают через cookie-сессию Mattermost.",
    "version": "0.1.0"
  },
  "servers": [{ "url": "/" }],
  "security": [{ "automationApiKey": [] }],
...
  "paths": {
    "/api/cron/scenarios": {
      "get": {
        "tags": ["Автоматизация"],
        "summary": "Актуальные сценарии поиска",
        "description": "Возвращает только активные (не архивные) сценарии поиска — «фильтры», которые автоматизация должна выполнять. Сценарий: ключевые слова, источники/ссылки, позитивные и негативные ключи, URL поиска (feedUrl) и/или регулярное выражение (regex) для фильтрации найденного. Приоритет: если задан feedUrl — автоматизация обращается к этому URL как есть (Яндекс, Google, RSS/API); иначе формирует запрос из keywords/sources. Найденные материалы фильтруются по regex (заголовок+описание), затем размещаются через POST /api/cron/mentions.",
        "security": [{ "automationApiKey": [] }],
        "responses": {
          "200": {
            "description": "Список активных сценариев",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScenariosResponse" },
                "examples": {
                  "scenarioWithFeedUrl": {
                    "summary": "Сценарий с URL поиска Яндекса (ежесуточно)",
                    "value": {
                      "boardId": "team",
                      "scenarios": [
                        {
                          "id": 1,
                          "projectId": "project-42",
                          "name": "Клеверспорт — мониторинг",
                          "keywords": ["КЛЕВЕРСПОРТ"],
                          "sources": [],
                          "negativeKeywords": ["суд", "штраф"],
                          "positiveKeywords": ["рост", "лидер"],
                          "regex": "(?:КЛЕВЕРСПОРТ|Клевер[- ]?Спорт)",
                          "feedUrl": "https://yandex.ru/search/?text=КЛЕВЕРСПОРТ&lr=213&within=1"
                        }
                      ]
                    }
                  },
                  "scenarioWithoutFeed": {
                    "summary": "Сценарий без feedUrl (автоматизация сама формирует запрос из keywords/sources)",
                    "value": {
                      "boardId": "team",
                      "scenarios": [
                        {
                          "id": 2,
                          "projectId": "project-7",
                          "name": "Имя директора — упоминания",
                          "keywords": ["Иванов И.И.", "гендиректор"],
                          "sources": [],
                          "negativeKeywords": [],
                          "positiveKeywords": [],
                          "regex": "",
                          "feedUrl": ""
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/DbError" }
        }
      }
    },
    "/api/cron/mentions": {
      "post": {
        "tags": ["Автоматизация"],
        "summary": "Разместить найденные материалы",
        "description": "Размещает найденные автоматизацией материалы в БД. Заголовок и краткая информация — отдельные поля (`title`, `comment`). Идемпотентно: при повторе строки с тем же (projectId, url) пропускаются. `projectId` можно задать один раз на уровне body — унаследуется всеми items.",
        "security": [{ "automationApiKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ImportMentionsRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Результат импорта",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImportMentionsResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/SearchFailed" }
        }
      }
    },
    "/api/cron/search-news": {
      "post": {
        "tags": ["Автоматизация"],
        "summary": "Выполнить сценарии поиска (сборка новостей на сервере)",
        "description": "Запускает все активные сценарии поиска за последние N дней и сохраняет совпадения в БД (source_type='auto_search'). Используется, когда поиск выполняет само приложение; обычно автоматизация сама ищет в Google News RSS и отдаёт результат через POST /api/cron/mentions.",
        "security": [{ "automationApiKey": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SearchNewsRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Сводка по сценариям",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SearchNewsResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/SearchFailed" }
        }
      }
    },
    "/api/team/login": {
      "post": {
        "tags": ["Кабинет команды"],
        "summary": "Вход в кабинет",
        "description": "Авторизация по логину/паролю Mattermost. Устанавливает cookie-сессию. Или с PAT-токеном (login_id + password с паспортом) — зависит от настроек сервера.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MeResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "description": "Неверный логин или пароль" }
        }
      }
    },
...
    "/api/team/me": {
      "get": {
        "tags": ["Кабинет команды"],
        "summary": "Текущий пользователь",
        "responses": {
          "200": {
            "description": "Данные пользователя и доступы",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MeResponse" } } }
          },
          "401": { "description": "Нет сессии" }
        }
      }
    },
    "/api/team/projects": {
      "get": {
        "tags": ["Кабинет команды"],
        "summary": "Список проектов (без архивных)",
        "responses": {
          "200": { "description": "Проекты борда", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectsResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/MattermostError" }
        }
      }
    },
    "/api/team/tasks": {
      "get": {
        "tags": ["Кабинет команды"],
        "summary": "Задачи команды",
        "description": "Задачи и факты публикаций. Параметры: project (id проекта), fact=1 (только факты публикаций), status (подстрока статуса).",
        "parameters": [
          { "name": "project", "in": "query", "schema": { "type": "string" }, "description": "ID проекта. Пусто — все проекты." },
          { "name": "fact", "in": "query", "schema": { "type": "string", "enum": ["0", "1"] }, "description": "Если 1 — только опубликованные факты." },
          { "name": "status", "in": "query", "schema": { "type": "string" }, "description": "Фильтр по статусу (подстрока)." }
        ],
        "responses": {
          "200": { "description": "Задачи", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TasksResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/MattermostError" }
        }
      }
    },
    "/api/team/search-scenarios": {
      "get": {
        "tags": ["Кабинет команды"],
        "summary": "Сценарии поиска (все)",
        "responses": {
          "200": { "description": "Сценарии", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScenariosResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/DbError" }
        }
      },
      "post": {
        "tags": ["Кабинет команды"],
        "summary": "Создать сценарий поиска",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScenarioInput" } } }
        },
        "responses": {
          "200": { "description": "Созданный сценарий", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScenarioResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/DbError" }
        }
      }
    },
    "/api/team/search-scenarios/{id}": {
      "put": {
        "tags": ["Кабинет команды"],
        "summary": "Обновить сценарий поиска",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScenarioInput" } } }
        },
        "responses": {
          "200": { "description": "Обновлённый сценарий", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScenarioResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "Сценарий не найден" },
          "502": { "$ref": "#/components/responses/DbError" }
        }
      },
      "delete": {
        "tags": ["Кабинет команды"],
        "summary": "Удалить сценарий поиска",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Удалён", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "Сценарий не найден" },
          "502": { "$ref": "#/components/responses/DbError" }
        }
      }
    },
    "/api/team/mentions": {
      "get": {
        "tags": ["Кабинет команды"],
        "summary": "Упоминания проекта (вкладка WEB)",
        "parameters": [{ "name": "project", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Список упоминаний", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MentionsResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/DbError" }
        }
      },
      "post": {
        "tags": ["Кабинет команды"],
        "summary": "Создать упоминание вручную",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MentionInput" } } }
        },
        "responses": {
          "200": { "description": "Созданное упоминание", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MentionResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/DbError" }
        }
      }
    },
...
    "/api/team/mentions/stats": {
      "get": {
        "tags": ["Кабинет команды"],
        "summary": "Помесячная статистика упоминаний",
        "parameters": [{ "name": "project", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Статистика", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MentionsStatsResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "502": { "$ref": "#/components/responses/DbError" }
        }
      }
    },
    "/api/links/{token}": {
      "get": {
        "tags": ["Клиентский кабинет"],
        "summary": "По токену ссылки клиента — имя проекта",
        "parameters": [{ "name": "token", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "Проект",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "boardId": { "type": "string" }, "projectId": { "type": "string" }, "name": { "type": "string" } } } } }
          },
          "404": { "description": "Ссылка недействительна" },
          "502": { "$ref": "#/components/responses/MattermostError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "automationApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Automation-Api-Key"
      }
    },
    "responses": {
      "Unauthorized": { "description": "Нет авторизации (неверный ключ автоматизации или отсутствует сессия)" },
...
      "DbError": { "description": "Ошибка базы данных", "content": { "application/json": { "schema": { "type": "object", "properties": { "error": { "type": "string" }, "message": { "type": "string" } } } } } },
...
    },
    "schemas": {
...
      "ScenariosResponse": {
        "type": "object",
        "properties": {
          "boardId": { "type": "string", "description": "Только в /api/cron/scenarios" },
          "scenarios": { "type": "array", "items": { "$ref": "#/components/schemas/Scenario" } }
        }
      },
...
      "ImportMentionsRequest": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "minItems": 1,
            "maxItems": 500,
            "items": { "$ref": "#/components/schemas/ImportItem" }
          },
          "providerName": { "type": "string", "default": "n8n", "description": "Кто разместил (visible createdBy)" },
          "projectId": { "type": "string", "description": "Общий проект для всех items" }
        }
      },
...
    }
  }
}